requirements.txt
openai==1.12.0
Pillow==10.2.0
streamlit==1.31.0
python-slugify==8.0.4
emoji==2.10.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.114.0
